feat: set up nav graph for app

+++ shared/src/commonMain/kotlin/org/ic/life/main/data/database/ic/ICEntity.kt
@@ -7,7 +7,7 @@ import org.ic.life.main.core.configs.AppConfig
 @Entity(tableName = AppConfig.IC_TABLE_NAME)
 data class ICEntity(
     @PrimaryKey(autoGenerate = true)
-    val uuid: Long,
+    val uuid: Int,
     val name: String,
     val address: String,
 )
